restructure recycle path to get both serach by geo-location and user input working
